pyproject.toml
[build-system]
requires = ["setuptools>=62.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["fluoraapi*"]

[project]
name = "fluoraapi"
version="0.0.1"
description="Python library for controlling Fluora LED plant"
readme="README.rst"
requires-python=">=3.10"
license = {file = "LICENSE.txt"}
authors = [
  {name = "drewancil", email = "andrew@crampedcell.com"},
]
keywords = ["fluora","fluoraapi"]
classifiers=[
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'License :: Freely Distributable',
    'Programming Language :: Python :: 3'
]
dependencies = ["python-osc", "python-box"]

[project.urls]
Repository = "https://github.com/drewancil/fluoraapi"

